refactor and added decryption
decryption uses pretty much the exact same logic as encryption, so we
might consider a function that does the basic stuff. The main difference
is how the locs are processed.

This may be worth looking into, to prevent multiple allocations of the
hashmap. (Or we could just use free on it)